requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pytest==7.4.3
black==23.11.0
flake8==6.1.0
openai==1.3.7
pytz==2023.3 
python-dotenv==1.0.0
psycopg2-binary==2.9.9

